forge fmt

+++ src/interfaces/ITimelock.sol
@@ -4,27 +4,15 @@ pragma solidity ^0.8.12;
 interface ITimelock {
     function queuedTransactions(bytes32) external view returns (bool);
 
-    function queueTransaction(
-        address target,
-        uint value,
-        string memory signature,
-        bytes memory data,
-        uint eta
-    ) external returns (bytes32);
+    function queueTransaction(address target, uint256 value, string memory signature, bytes memory data, uint256 eta)
+        external
+        returns (bytes32);
 
-    function executeTransaction(
-        address target,
-        uint value,
-        string memory signature,
-        bytes memory data,
-        uint eta
-    ) external payable returns (bytes memory);
+    function executeTransaction(address target, uint256 value, string memory signature, bytes memory data, uint256 eta)
+        external
+        payable
+        returns (bytes memory);
 
-    function cancelTransaction(
-        address target,
-        uint value,
-        string memory signature,
-        bytes memory data,
-        uint eta
-    ) external;
+    function cancelTransaction(address target, uint256 value, string memory signature, bytes memory data, uint256 eta)
+        external;
 }
